+++ networking/ilb-next-hop/gateways.tf
@@ -33,15 +33,13 @@ module "gw" {
       network    = module.vpc-left.self_link
       subnetwork = values(module.vpc-left.subnet_self_links)[0],
       nat        = false,
-      addresses  = null,
-      alias_ips  = null
+      addresses  = null
     },
     {
       network    = module.vpc-right.self_link
       subnetwork = values(module.vpc-right.subnet_self_links)[0],
       nat        = false,
-      addresses  = null,
-      alias_ips  = null
+      addresses  = null
     }
   ]
   tags           = ["ssh"]