+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- new `cloud-run` module
 
 ## [8.0.0] - 2021-10-21